backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.12.5
pydantic-settings==2.1.0
email-validator>=2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.27.2
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
ollama==0.6.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4